+++ CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 - Added commenter support
+- Added formatting and code style settings
 
 ## 0.6.0-beta (2026-02-07)
 - Added parameter info support