+++ CHANGELOG.md
@@ -4,12 +4,12 @@ All notable changes to this project will be documented in this file.
 
 ---
 
-## [Unreleased]
+## 0.7.0-beta (2026-02-27)
 - Added commenter support
 - Added formatting and code style settings
 - Added uniform hint documentation support
 - Added inlay parameter hints feature
-- Added surrounding support
+- Added surround with support
 
 ## 0.6.0-beta (2026-02-07)
 - Added parameter info support